rename: 파일명 수정(CloseBtn -> ChallengeDeleteBtn)
